refactor: create showcase for select field
